Make memory usage tracking thread-safe
Fixes #296. This keeps track of allocation sizes by adding a header to
every allocation holding the size. The difference between the memory
usage we've last reported to V8 and the current memory usage is tracked
in `xml_memory_diff`, and protected by `xml_memory_mutex`. It is
frequently synchronised by `xml_memory_cb`, which is called on the main
thread through `xml_memory_handle`.